add library variant
