+++ PermissionExample/app/src/main/AndroidManifest.xml
@@ -5,12 +5,13 @@
     <uses-permission android:name="android.permission.WRITE_CONTACTS"/>
 
     <application
+        android:name=".App"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".MainActivity">
+        <activity android:name=".activities.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -18,8 +19,8 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".LibraryVersionActivity"/>
-        <activity android:name=".NativeVersionActivity"/>
+        <activity android:name=".activities.LibraryVersionActivity"/>
+        <activity android:name=".activities.NativeVersionActivity"/>
     </application>
 
 </manifest>
