--- PermissionExample/app/src/main/java/fs/com/permissionexample/LibraryVersionActivity.java
@@ -1,19 +1,0 @@
-package fs.com.permissionexample;
-
-import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-
-import butterknife.ButterKnife;
-
-/**
- * Created by Ramil on 09/02/16.
- */
-public class LibraryVersionActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-        ButterKnife.bind(this);
-    }
-}
