+++ PermissionExample/app/src/main/java/fs/com/permissionexample/activities/MainActivity.java
@@ -1,4 +1,4 @@
-package fs.com.permissionexample;
+package fs.com.permissionexample.activities;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
@@ -6,6 +6,7 @@ import android.os.Bundle;
 
 import butterknife.ButterKnife;
 import butterknife.OnClick;
+import fs.com.permissionexample.R;
 
 public class MainActivity extends AppCompatActivity {
