+++ PermissionExample/app/src/main/res/layout/activity_main.xml
@@ -8,7 +8,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="fs.com.permissionexample.MainActivity">
+    tools:context=".activities.MainActivity">
 
     <Button
         android:id="@+id/btn_native"
